Add a searcher for monitoring dashboards
This allows something like `gcp monitoring dashboards my_service`,
which then opens that particular dashboard in the browser. I wanted this
because I hate having to do a lot of clicking to find the dashboard I want to
look at when an alert triggers or if I need some quick info.

If you hit <Enter> before searching it just takes you to the list of
dashboards, which also seems helpful.